Moved charts to dash for better filter performance (#1)
Moved charts to dash for better performance. Dashboard now shows PNL Graph and indicators for PCR % und Win rate %.
Plotly charts filtering had too much limitations.
Removed some unsed files and functions and reworked the transform function.
The functions now used fifoPnlRealized if the trade is settled and mtmpnl if the trade is closed but not yet settled. The opening date/time is now extracted to the closed position to group combo trades. This was necessary for the correct calcutions for win rate.